Source/KhoaLuanBE/requirements.txt
pydantic~=2.8.2
uvicorn~=0.30.1
fastapi~=0.111.0
SQLAlchemy~=2.0.31
databases~=0.9.0
tensorflow~=2.13.0
matplotlib~=3.7.5
numpy~=1.24.3
pillow~=10.4.0